Senior-level code review round 2: numpy perf, thread safety, code quality

+++ youtubedl_saver.py
@@ -16,7 +16,7 @@ def save_file(url, outtmpl=None):
         try:
             info = ydl.extract_info(url=url, download=True)
             fps = info.get('fps', 30)
-            total_frames = fps * info['duration']
+            total_frames = int(fps * info['duration'])
             duration = info['duration']
         except Exception as e:
             print(f"Error downloading video: {e}")
